ODK-200 Additional server side check to prevent users from changing OfficeId after creating an office.

+++ src/main/java/org/benetech/controller/UserAdminController.java
@@ -3,6 +3,7 @@ package org.benetech.controller;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.commons.codec.binary.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.benetech.ajax.AjaxFormResponse;
@@ -20,6 +21,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -72,12 +74,18 @@ public class UserAdminController {
 
   @Secured({"ROLE_SITE_ACCESS_ADMIN"})
   @PostMapping("/admin/users/delete")
-  public String deleteUser(@ModelAttribute("user") UserEntityForm user, Model model) {
+  public String deleteUser(@ModelAttribute("user") UserEntityForm user, Model model, Authentication authentication) {
     OdkClient odkClient = odkClientFactory.getOdkClient();
-    HttpStatus status = odkClient.deleteUser(user.getUsername());
+    if (StringUtils.equals(authentication.getName(), user.getUsername())) {
+      model.addAttribute("msg", "This client does not support deleting the currently logged-in user.");
+      model.addAttribute("css", "danger");
+    } else {
+      HttpStatus status = odkClient.deleteUser(user.getUsername());
+      model.addAttribute("msg", "User " + user.getUsername() + " has been deleted.");
+      model.addAttribute("css", "info");
+    }
     populateDefaultModel(model);
-    model.addAttribute("msg", "User " + user.getUsername() + " has been deleted.");
-    model.addAttribute("css", "info");
+
     return "admin_user_grid";
   }
